requirements.txt
# FortiFun API Server with WebSocket Support
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0

# WebSocket Support
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0

# Database
psycopg2>=2.9.0
dj-database-url>=1.0.0

# File handling
Pillow>=10.0.0
django-storages>=1.14.2
boto3>=1.34.0

# Production
gunicorn>=20.1.0
whitenoise>=6.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT>=2.8.0
drf-nested-routers>=0.93.4

# Health check
django-health-check>=3.17.0
